feat: Добавить конфик файл
